Website Base
I have built a frame for the website.

There are 2 pages:

Home page - All that needs to be done is write the description of the project and what was done.

Reports page - I have made a sample entry. Once we have all of the visualizations and reports then we just need to replicated what has been done for the sample and it should be ready to go. I have used my group as an example.